chore(sp314-dsp): resolve F-032 — 5 targeted clippy findings plus 10 related lint fixes
The 5 findings from F-032 (collapsible_if, excessive_precision,
cloned_ref_to_slice_refs, legacy_numeric_constants, useless_vec)
plus 10 additional pre-existing findings surfaced by the same
correct -D warnings mirror across the crate's full test suite
(manual_range_contains, div_ceil, an assertions_on_constants case,
and one more excessive_precision instance) — all mechanical,
clippy-suggested rewrites, zero behavior change.

Two categories got extra verification before acceptance rather than
blind trust in "tests still pass":

- harmonic_contract.rs's assert! -> const { assert!(...) } is a
  clippy-suggested compile-time-check conversion (assertions_on_
  constants lint) — confirmed equivalent since GLOBAL_K_HARMONIC is
  a compile-time constant; zero runtime behavior change.
- All truncated floating-point literals (harmonic_contract.rs's
  calibration constants — PAD_LINEAR, GLOBAL_K_HARMONIC, and 6
  PER_DRIVE values — plus pre_analysis_contract.rs's two_pi fixture
  constant) were verified via explicit to_bits() round-trip: each
  truncated literal parses back to the IDENTICAL f32 bit pattern as
  the original high-precision literal. No calibration reference or
  test fixture drifted by even one bit.

Full workspace clippy (-D warnings, three project-allowed lint
exceptions) is now clean across sp314-dsp. All existing tests pass
unchanged.

Resolves F-032.

+++ lineos/m1/sp314-dsp/tests/limiter_contract.rs
@@ -32,7 +32,7 @@ fn limiter_ring_buffer_wraps_correctly() {
         let val = (i as f32) + 1.0;
         let delayed = rb.push_and_pop(val);
         if i >= 240 {
-            assert_eq!(delayed, val - 240 as f32);
+            assert_eq!(delayed, val - 240_f32);
         } else {
             assert_eq!(delayed, 0.0);
         }
